+++ Car.cpp
@@ -23,7 +23,7 @@ int Car::getDoors(){return numOfDoors;}
 
 
 
-double Car::costPerDay(){
+int Car::costPerDay(){
   int costAsPence = 2500 - (age*150) - (numOfDoors*200);
   return penceToPounds(perDayCostCap(costAsPence));
